refactor: 이름 변경, 폴더 구조 변경
