+++ src/pages/games/AppleGame/game/model/snackGame.ts
@@ -6,7 +6,6 @@ export class SnackGame {
   private row: number;
   private column: number;
   private borderOffset = 10;
-  private selectStarted = false;
   private selectedSnacks: Snack[] = [];
 
   constructor({
